+++ inst/NEWS.md
@@ -6,6 +6,8 @@
   * BUG FIX `project_to_curve()`: set dimension names of outputted `s` correctly.
   
   * DOCUMENTATION: Added `cran-comments.md` and `revdep` to repository.
+  
+  * MINOR CHANGE: Removed `adjust_range()`; use `grDevices::extendrange()` instead.
 
 # princurve 2.0.4 (2018-07-09)